scripts: split the three new gates under the 500-line ratchet; a push step is paired with the git add before it; measured test count
- check_gate_reachability.py, check_timeout_nesting.py and
  check_api_reference.py were 987, 955 and 595 lines; each is now a thin
  entry point over a package under scripts/lib/ (gate_reachability/,
  timeout_nesting/, api_reference_roots.py), every file under 500 lines.
  Behaviour proved unchanged: default, --explain and --strict outputs
  byte-identical to the pre-split scripts, same exit codes, from three
  working directories, and every injection re-run against both versions
  with identical output. scripts/check_file_lengths.sh: 82 of 498
  tracked files over the limit, all recorded, exit 0.
- gate_reachability: benchmarks.yml now commits, grades and pushes in
  three steps, and a push step whose only command is `git push` stages
  nothing itself, so the pushed files fell back to `**` and deleting the
  prose gate went unnoticed — the B13 "done when" case. The push is now
  paired with the `git add` of the job's earlier steps. Verified: the
  current tree reports 0 findings; deleting the prose step reports
  BYPASSED for benchmarks.md; removing ci.yml's push trigger reports
  UNREACHABLE.
- SPEC_COMPLIANCE.md's "~2,500 tests" is the measured figure: 3,198
  passed, 0 failed, 175 ignored across 103 binaries (`cargo test
  --workspace --all-features --no-fail-fast`, 2026-09-10).
