+++ scripts/check_gate_reachability.py
@@ -33,65 +33,18 @@ Two instances is a class. This is the checker for the class: it reads the real
 asserts that each gate is reachable from every event that can change its
 inputs.
 
-The model, in five lines:
+Where the checker lives
+-----------------------
+This file is the entry point only. The checker is the package
+`scripts/lib/gate_reachability/`; its `__init__.py` states the model (gates,
+events, rules R1/R2/R3) and what the checker does and does not prove, and each
+refinement is documented next to the code that implements it. The two things
+a maintainer is most likely to edit:
 
-  1. A *gate* is a `run:` step. Its *inputs* are the tracked files it reads,
-     taken from an explicit table (`GATE_INPUTS`) for the repository's script
-     gates, derived from `-p <package>` for cargo commands, and "unknown" for
-     anything else. Unknown is never rounded to "reads nothing".
-  2. An *event that changes inputs on main* is a `push` to `main` — a merge, a
-     direct push, or a workflow's own push. `pull_request` runs see a proposed
-     tree, not `main`. `schedule` sees `main`, late. `workflow_dispatch` sees
-     nothing unless somebody clicks.
-  3. R1 — every gate-bearing job must be reachable from `push` to `main` or
-     from `schedule`, after its own `if:` is applied. A job that its `if:`
-     confines to `pull_request` is the PR half of a pair and needs a sibling
-     that covers `main`. A tag-only workflow is reachable from the tag push.
-  4. R2 — when the reaching `push` trigger carries a `paths:` filter, the
-     filter must cover the gate's inputs, unless another main-reachable job
-     runs the same command without that gap.
-  5. R3 — every `git push` a workflow performs with `GITHUB_TOKEN` creates no
-     events, so for every file that push changes, some gate that reads that
-     file must run *earlier in the same job* (or in a job it `needs`). If no
-     gate anywhere reads the file there is nothing to escape; if gates read it
-     elsewhere and none runs on the push path, that is the Q6 defect.
-
-What this checks, and what it does not
---------------------------------------
-It checks the three rules above over every `.github/workflows/*.yml`, and it
-exits 1 on the first tree that breaks any of them. It has no allowlist in use:
-`ALLOWED` exists so that a finding accepted on purpose can be recorded under
-the key the finding prints, with its reason, and every accepted entry is
-printed as ALLOWED on every run.
-
-Two refinements the two instances forced. A step that decides with a bare
-`exit 1` and no recognisable command (dco.yml's shape) is still a gate; the
-first version of this script could not see it, which is the failure mode it
-exists to prevent. And a workflow's push adds a *commit* as well as files, so
-a gate that reads commit metadata (`git log`) reads every push — that is how
-the two instances interlock: the benchmark bot's commit is exactly what the
-DCO gate would reject, and neither gate ever sees it. Tag-only workflows are
-left out of R3's reader set: a push to `main` is not a tag push, and their
-gates see the tree only when a tag is cut.
-
-It does not prove a gate can fail (the two provers do that), does not evaluate
-branch protection (a `pull_request`-only gate is fine *only* if nothing but
-merges ever reach `main`, and this repository's own bot disproves that), and
-does not read scripts to discover their inputs — the table is hand-written and
-says so, because a heuristic that scans scripts for path literals returned
-mostly self-references and comment text when tried. A gate missing from the
-table is reported as "inputs unknown" in `--explain`, never counted as
-coverage for R3, and skipped by R2. That is the conservative direction for a
-reachability checker: an unknown gate can never make another gate look
-covered.
-
-`if:` conditions are evaluated for `github.event_name` comparisons, `always()`,
-`||`, `&&` and `!`; any other atom (`needs.*.outputs.*`, `inputs.*`) is taken
-as true, and the condition is printed verbatim in `--explain` so that
-assumption can be audited. Path coverage under R2 is tested by matching one
-representative path per input glob against the trigger's `paths:` patterns
-with GitHub's filter semantics (`*` stops at `/`, `**` does not), which is
-exact for the prefix-shaped patterns this repository uses.
+  * `GATE_INPUTS` — the hand-written table of tracked files each script gate
+    reads — is in `scripts/lib/gate_reachability/inputs.py`.
+  * `ALLOWED` — findings accepted on purpose, empty by design — is in
+    `scripts/lib/gate_reachability/rules.py`.
 
 Usage
 -----
@@ -111,831 +64,19 @@ the inputs (no workflows directory, unparsable YAML, PyYAML missing).
 from __future__ import annotations
 
 import argparse
-import re
-import shlex
 import sys
-import tomllib
-from dataclasses import dataclass, field
 from pathlib import Path
 
-try:
-    import yaml
-except ModuleNotFoundError:  # pragma: no cover - environment problem, not logic
-    sys.exit(
-        "error: PyYAML is required.\n"
-        "  pip install pyyaml   (GitHub's ubuntu runners ship it preinstalled)\n"
-        "A checker that hand-parses `on:` blocks is a checker with a second\n"
-        "YAML implementation, and that is how it starts skipping the workflow\n"
-        "whose trigger it cannot read."
-    )
-
-REPO = Path(__file__).resolve().parent.parent
-WORKFLOWS = REPO / ".github" / "workflows"
-DEFAULT_BRANCH = "main"
-
-# Events that fire when `main` changes. `schedule` is late but it does see
-# `main`; `pull_request` and `workflow_dispatch` never do on their own.
-MAIN_EVENTS = ("push", "schedule")
-ALL_EVENTS = ("push", "pull_request", "schedule", "workflow_dispatch")
-
-# Findings accepted on purpose, keyed exactly as the finding prints its key
-# (`unreachable:dco.yml`, `filter-gap:ci.yml:fmt:step[6]`,
-# `bypassed:benchmarks.yml:bench:step[22]:(commit history)`) -> reason.
-# Empty by design. An accepted finding is printed as ALLOWED on every run with
-# its reason, so nothing it accepts is silent. Add one only with a reason a
-# reviewer can check against the workflow it names.
-ALLOWED: dict[str, str] = {}
-
-EXPR = re.compile(r"\$\{\{\s*(.+?)\s*\}\}")
-
-# ── Gate inputs ──────────────────────────────────────────────────────────────
-#
-# Tracked files each script gate reads, keyed by a regex over the command. The
-# table is hand-written from the scripts' own file access; it is what R2 and R3
-# reason over, so a wrong entry is a wrong verdict. An empty list means the
-# gate reads only run-time output (criterion reports, conformance reports) and
-# has no tracked inputs — distinct from absent, which means unknown.
-GATE_INPUTS: list[tuple[str, list[str]]] = [
-    # The DCO check over a commit range; benchmarks.yml runs it over the
-    # commit it is about to push, which is what covers a GITHUB_TOKEN push.
-    (r"check_dco\.sh", ["(commit history)"]),
-    # The Q6 instance: reads the generated page benchmarks.yml pushes.
-    (r"check_benchmark_prose\.sh", ["book/src/reference/benchmarks.md"]),
-    # Counts fences in every book page and requires each to be registered.
-    (r"check_book_code\.sh", ["book/src/**/*.md", "book-tests/src/lib.rs",
-                              ".book-ignore-baseline"]),
-    (r"check_proto_copies\.sh", ["**/*.proto"]),
-    (r"check_file_lengths\.sh", ["**/*.rs", "**/*.sh", "**/*.py"]),
-    (r"check_mutation_scope\.sh", [".github/workflows/mutants.yml",
-                                   "crates/*/src/**/*.rs"]),
-    (r"gen_sitemap\.py", ["book/src/SUMMARY.md", "book/static/sitemap.xml",
-                          "book/static/robots.txt"]),
-    (r"check_api_reference\.py", ["book/src/reference/api-reference.md",
-                                  "crates/**/*.rs"]),
-    (r"check_otel_metrics_coverage\.py", ["crates/**/*.rs"]),
-    (r"check_package_excludes\.py", ["**/Cargo.toml", ".github/workflows/ci.yml",
-                                     ".github/workflows/release.yml",
-                                     "RELEASING.md"]),
-    (r"prove_workflow_gates_fail\.py", [".github/workflows/**", "scripts/**",
-                                        "tck/scripts/**"]),
-    (r"check_block_scalars\.py", ["scripts/lib/ci_gates.sh"]),
-    (r"check_cancellation_release\.py", ["crates/*/src/**/*.rs",
-                                         "bindings/*/src/**/*.rs"]),
-    (r"check_doc_escapes\.py", ["crates/*/src/**/*.rs", "bindings/*/src/**/*.rs",
-                                "examples/*/src/**/*.rs"]),
-    (r"check_panic_paths\.py", ["crates/*/src/**/*.rs"]),
-    (r"check_codecov_ignores\.py", ["codecov.yml"]),
-    (r"check_provenance_manifest\.py", ["**"]),
-    (r"check_slimrpc_spec\.sh", ["bindings/a2a-protocol-slimrpc/**"]),
-    (r"check_method_denominator\.py", ["crates/a2a-protocol-types/src/method.rs",
-                                       "bindings/a2a-protocol-slimrpc/**"]),
-    (r"package_binding\.py", ["bindings/a2a-protocol-slimrpc/**"]),
-    (r"grpc_wire_compat\.py", ["tck/**", "crates/a2a-protocol-types/**"]),
-    (r"python_client_vs_rust\.py|itk_traversal_selftest\.py|inspector_card_check\.py",
-     ["itk/**"]),
-    (r"^\./target/release/a2a-tck(-sut)?\b", ["tck/**", "crates/**"]),
-    # mdBook parses the summary and the markdown; other files are copied as-is.
-    (r"^mdbook build", ["book/book.toml", "book/src/**/*.md"]),
-    # Run-time output only: criterion's target/criterion, the TCK's report.
-    (r"check_streaming_linearity\.py", []),
-    (r"check_regression\.py", []),
-    (r"check_conformance\.py", []),
-]
-
-# Any cargo command reads the workspace's sources and manifests.
-RUST_TREE = ["**/*.rs", "**/Cargo.toml", "**/Cargo.lock", "**/*.proto"]
-# `book-tests/src/lib.rs` is `#[doc = include_str!(...)]` of every book page,
-# so building its docs or running its doctests reads the book. Measured: the
-# generated benchmarks page is registered there (book-tests/src/lib.rs:197).
-CARGO_READS_BOOK = ("test", "doc", "llvm-cov")
-CARGO_GATES = ("fmt", "clippy", "test", "build", "doc", "package", "hack", "run",
-               "bench", "llvm-cov", "semver-checks", "deny", "mutants", "nextest")
-
-# Commands that set up a step rather than decide it. Filtered so `--explain`
-# lists gates, not plumbing; nothing here can be an input reader in R3 either.
-NOT_A_GATE = re.compile(
-    r"^(set|echo|printf|mkdir|cd|export|source|\.|git|pip|pip3|uv|rustup|sleep|cat|cp|ln|"
-    r"mv|rm|curl|wget|tee|chmod|true|false|exit|count|if|then|else|elif|fi|for|while|do|"
-    r"done|case|esac|break|continue|return|wait|kill|test|\[|\[\[|\{|\}|\(|python3? --version|"
-    r"python3? -m pip|cargo install|cargo metadata|diff|find|sort|head|tail|grep|sed|awk|"
-    r"jq|xargs|tar|unzip|gzip|sha256sum|gh|date|read|local|declare|trap|shift|seq|basename|"
-    r"dirname|tr|ls|shopt|pwd|which|command|type|nproc|uname|env|id|whoami|hostname|"
-    r"realpath|touch|wc|cut|GHEXPR|[0-9]+|\(\(|!|>|<)(\s|$)"
-)
-
-# A step that decides with a bare `exit 1` — dco.yml's shape — has no command
-# to look up but is a gate all the same. Same pattern as the prover's
-# EXPLICIT_FAIL: any non-zero exit, anywhere on the line.
-EXPLICIT_FAIL = re.compile(r"\bexit\s+(?:[1-9][0-9]*|\"?\$)")
-
-# Not a tracked file: a gate that reads commit metadata reads every commit,
-# and a workflow's own push adds one.
-HISTORY = "(commit history)"
-
-# Actions that commit and push to the repository. `docker/build-push-action`
-# pushes to a registry, not to git, and must not match.
-PUSH_ACTIONS = re.compile(r"git-auto-commit-action|add-and-commit|github-push-action")
-
-
-# ── Workflow model ───────────────────────────────────────────────────────────
-
-
-@dataclass
-class Trigger:
-    event: str
-    branches: list[str] | None = None
-    tags: list[str] | None = None
-    paths: list[str] | None = None
-    paths_ignore: list[str] | None = None
-
-    @property
-    def tag_only(self) -> bool:
-        return self.event == "push" and self.tags is not None and self.branches is None
-
-    def covers_branch(self, branch: str) -> bool:
-        if self.tag_only:
-            return False
-        if self.branches is None:
-            return True
-        return any(gh_match(p, branch) for p in self.branches)
-
-    def describe(self) -> str:
-        bits = [self.event]
-        if self.branches is not None:
-            bits.append("branches=" + ",".join(self.branches))
-        if self.tags is not None:
-            bits.append("tags=" + ",".join(self.tags))
-        if self.paths is not None:
-            bits.append("paths=" + ",".join(self.paths))
-        if self.paths_ignore is not None:
-            bits.append("paths-ignore=" + ",".join(self.paths_ignore))
-        return " ".join(bits)
-
-
-@dataclass
-class Gate:
-    workflow: str
-    job: str
-    index: int
-    name: str
-    commands: list[str]
-    inputs: list[str] | None  # None = unknown
-    condition: str | None
-
-    @property
-    def where(self) -> str:
-        return f"{self.workflow}:{self.job}:step[{self.index}] {self.name!r}"
-
-    def reads(self, path: str) -> bool:
-        if self.inputs is None:
-            return False
-        if path == HISTORY or HISTORY in self.inputs:
-            return path in self.inputs
-        return any(globs_intersect(g, path) for g in self.inputs)
-
-
-@dataclass
-class Push:
-    workflow: str
-    job: str
-    index: int
-    name: str
-    files: list[str]
-    token: str  # "GITHUB_TOKEN" or the secret's name
-    line: str
-
-    @property
-    def where(self) -> str:
-        return f"{self.workflow}:{self.job}:step[{self.index}] {self.name!r}"
-
-    @property
-    def creates_events(self) -> bool:
-        return self.token != "GITHUB_TOKEN"
-
-
-@dataclass
-class Job:
-    workflow: str
-    name: str
-    condition: str | None
-    needs: list[str]
-    gates: list[Gate]
-    pushes: list[Push]
-    pr_range: bool  # reads a github.event.pull_request base..head range
-    runs_on: dict[str, bool | None] = field(default_factory=dict)
-
-
-@dataclass
-class Workflow:
-    file: str
-    triggers: list[Trigger]
-    jobs: dict[str, Job]
-
-    def trigger(self, event: str) -> Trigger | None:
-        return next((t for t in self.triggers if t.event == event), None)
-
-    @property
-    def tag_only(self) -> bool:
-        push = self.trigger("push")
-        return push is not None and push.tag_only
-
-    def main_events(self) -> list[str]:
-        """Events that fire for this workflow when `main` changes."""
-        out = []
-        push = self.trigger("push")
-        if push is not None and push.covers_branch(DEFAULT_BRANCH):
-            out.append("push")
-        if self.trigger("schedule") is not None:
-            out.append("schedule")
-        return out
-
-
-# ── Glob semantics (GitHub filter patterns) ──────────────────────────────────
-
-
-def gh_regex(pattern: str) -> re.Pattern[str]:
-    out, i = [], 0
-    while i < len(pattern):
-        c = pattern[i]
-        if pattern.startswith("**/", i):
-            out.append("(?:.*/)?")
-            i += 3
-        elif pattern.startswith("/**", i) and i + 3 == len(pattern):
-            out.append("(?:/.*)?")
-            i += 3
-        elif pattern.startswith("**", i):
-            out.append(".*")
-            i += 2
-        elif c == "*":
-            out.append("[^/]*")
-            i += 1
-        elif c == "?":
-            out.append("[^/]")
-            i += 1
-        else:
-            out.append(re.escape(c))
-            i += 1
-    return re.compile("^" + "".join(out) + "$")
-
-
-def gh_match(pattern: str, path: str) -> bool:
-    return gh_regex(pattern).match(path) is not None
-
-
-def representative(glob: str) -> str:
-    """One concrete path that the glob matches, for testing a filter against it."""
-    rep = glob.replace("**/", "d1/d2/")
-    if rep.endswith("/**"):
-        rep = rep[:-3] + "/d1/d2"
-    rep = rep.replace("**", "d1/d2").replace("*", "x").replace("?", "x")
-    return re.sub(r"\[([^\]])[^\]]*\]", r"\1", rep)
-
-
-def is_glob(s: str) -> bool:
-    return any(ch in s for ch in "*?[")
-
-
-def globs_intersect(a: str, b: str) -> bool:
-    if a == "**" or b == "**":
-        return True
-    if not is_glob(b):
-        return gh_match(a, b)
-    if not is_glob(a):
-        return gh_match(b, a)
-    return gh_match(a, representative(b)) or gh_match(b, representative(a))
-
-
-def covered_by_filter(glob: str, trigger: Trigger) -> bool:
-    rep = representative(glob)
-    if trigger.paths_ignore and any(gh_match(p, rep) for p in trigger.paths_ignore):
-        return False
-    if trigger.paths is None:
-        return True
-    return any(gh_match(p, rep) for p in trigger.paths)
-
-
-# ── `if:` evaluation ─────────────────────────────────────────────────────────
-
-
-def runs_on(condition: str | None, event: str) -> bool | None:
-    """Whether a condition holds for `github.event_name == event`.
-
-    Returns None when the expression has a shape this does not understand;
-    callers treat that as "may run" and `--explain` prints the text.
-    """
-    if condition is None:
-        return True
-    s = EXPR.sub(r"\1", str(condition)).strip()
-    if s in ("true", "True"):
-        return True
-    if s in ("false", "False"):
-        return False
-
-    def verdict(op: str, value: str) -> str:
-        return " True " if (op == "==") == (value == event) else " False "
-
-    s = re.sub(r"github\.event_name\s*(==|!=)\s*'([^']*)'",
-               lambda m: verdict(m.group(1), m.group(2)), s)
-    s = re.sub(r"'([^']*)'\s*(==|!=)\s*github\.event_name",
-               lambda m: verdict(m.group(2), m.group(1)), s)
-    # Status functions take their value on the normal path: nothing was
-    # cancelled, nothing failed. Any other call (`contains(a, 'b')`),
-    # comparison or bare atom is unknown here; assume it holds.
-    s = re.sub(r"\b(always|success)\(\)", " True ", s)
-    s = re.sub(r"\b(cancelled|failure)\(\)", " False ", s)
-    s = re.sub(r"\b[A-Za-z_]\w*\([^()]*\)", " True ", s)
-    s = re.sub(r"[A-Za-z_][\w.\-\[\]]*\s*(==|!=)\s*'[^']*'", " True ", s)
-    s = s.replace("||", " or ").replace("&&", " and ")
-    s = re.sub(r"!(?!=)", " not ", s)
-    s = re.sub(r"\b(?!(?:True|False|and|or|not)\b)[A-Za-z_][\w.\-\[\]]*", " True ", s)
-    if not re.fullmatch(r"[\sTrueFalseandornot()]+", s):
-        return None
-    try:
-        return bool(eval(s, {"__builtins__": {}}, {}))  # noqa: S307 - vetted charset
-    except SyntaxError:
-        return None
-
-
-# ── Shell reading ────────────────────────────────────────────────────────────
-
-HEREDOC = re.compile(r"<<-?\s*'?\"?([A-Za-z_][A-Za-z0-9_]*)")
-
-
-def shell_commands(body: str) -> list[str]:
-    """Simple commands in a `run:` body, skipping comments, heredocs, and
-    the arguments of echo/printf (which is where dco.yml and release.yml
-    print `git push` as advice rather than run it)."""
-    out: list[str] = []
-    terminator: str | None = None
-    pending = ""  # a command continued by `\` or by an open quote
-    body = EXPR.sub("GHEXPR", body)
-    body = re.sub(r"\$\(\([^)]*\)\)", "0", body)  # arithmetic is never a command
-    for raw in body.splitlines():
-        line = raw.rstrip()
-        if terminator is not None:
-            if line.strip() == terminator:
-                terminator = None
-            continue
-        stripped = line.strip()
-        if not pending and (not stripped or stripped.startswith("#")):
-            continue
-        if pending:
-            stripped = pending + "\n" + stripped
-            pending = ""
-        if stripped.endswith("\\"):
-            pending = stripped[:-1].rstrip()
-            continue
-        m = HEREDOC.search(stripped)
-        if m:
-            terminator = m.group(1)
-            stripped = stripped[: m.start()].rstrip()
-            if not stripped:
-                continue
-        try:
-            lexer = shlex.shlex(stripped, posix=True, punctuation_chars="|&;()")
-            lexer.whitespace_split = True
-            tokens = list(lexer)
-        except ValueError:  # an open quote: the command continues on the next line
-            if stripped.count("\n") < 20:
-                pending = stripped
-                continue
-            tokens = stripped.split()
-        cur: list[str] = []
-        for tok in tokens:
-            if tok in ("&&", "||", ";", "|", "&", "(", ")", ";;", "|&"):
-                if cur:
-                    out.append(" ".join(cur))
-                cur = []
-            else:
-                cur.append(tok.replace("\n", " "))
-        if cur:
-            out.append(" ".join(cur))
-    if pending:
-        out.append(pending)
-    return [c for c in out if c]
-
-
-def is_assignment(cmd: str) -> bool:
-    return re.match(r"^[A-Za-z_][A-Za-z0-9_]*(\[[^\]]*\])?[+]?=|^[A-Za-z_][A-Za-z0-9_]* \(\)", cmd) is not None
-
-
-def gate_commands(body: str) -> list[str]:
-    """The commands in a body that could carry a verdict."""
-    cmds = []
-    for c in shell_commands(body):
-        c = re.sub(r"^(?:[A-Za-z_][A-Za-z0-9_]*=\S*\s+)+", "", c)  # env prefix
-        if not c or is_assignment(c) or NOT_A_GATE.match(c):
-            continue
-        cmds.append(c)
-    return cmds
-
-
-def normalise(cmd: str) -> str:
-    cmd = re.sub(r"^(python3?|bash|sh)\s+", "", cmd)
-    return cmd[2:] if cmd.startswith("./") else cmd
-
-
-# ── Inputs ───────────────────────────────────────────────────────────────────
-
-
-def package_dirs() -> dict[str, str]:
-    """`[package] name` -> directory, from every tracked Cargo.toml."""
-    out: dict[str, str] = {}
-    for toml in REPO.rglob("Cargo.toml"):
-        if "target" in toml.parts or "node_modules" in toml.parts:
-            continue
-        try:
-            data = tomllib.loads(toml.read_text(encoding="utf-8"))
-        except (tomllib.TOMLDecodeError, OSError):
-            continue
-        name = (data.get("package") or {}).get("name")
-        if name:
-            rel = toml.parent.relative_to(REPO).as_posix()
-            out[name] = "" if rel == "." else rel
-    return out
-
-
-def cargo_inputs(cmd: str, packages: dict[str, str]) -> list[str] | None:
-    m = re.match(r"^cargo\s+(?:\+\S+\s+)?([a-z-]+)", cmd)
-    if not m or m.group(1) not in CARGO_GATES:
-        return None
-    sub = m.group(1)
-    named = re.findall(r"(?:-p|--package)\s+(\S+)", cmd)
-    if named and all(n in packages for n in named):
-        inputs = ["Cargo.toml", "Cargo.lock", "crates/**"]
-        for name in named:
-            d = packages[name]
-            inputs.append(f"{d}/**" if d else "**")
-    else:  # the workspace, or a package named by an expression (`-p GHEXPR`)
-        inputs = list(RUST_TREE)
-    if sub in CARGO_READS_BOOK:
-        inputs.append("book/src/**/*.md")
-    return inputs
-
-
-def inputs_for(commands: list[str], packages: dict[str, str]) -> list[str] | None:
-    """Union of known inputs; None if any command is unknown and none known."""
-    found: list[str] = []
-    known = False
-    for cmd in commands:
-        c = cargo_inputs(cmd, packages)
-        if c is None:
-            for pattern, ins in GATE_INPUTS:
-                if re.search(pattern, cmd):
-                    c = ins
-                    break
-        if c is not None:
-            known = True
-            for g in c:
-                if g not in found:
-                    found.append(g)
-    return found if known else None
-
-
-# ── Loading ──────────────────────────────────────────────────────────────────
-
-
-def as_list(v: object) -> list[str] | None:
-    if v is None:
-        return None
-    if isinstance(v, str):
-        return [v]
-    return [str(x) for x in v]
-
-
-def parse_triggers(on: object) -> list[Trigger]:
-    if isinstance(on, str):
-        return [Trigger(on)]
-    if isinstance(on, list):
-        return [Trigger(str(e)) for e in on]
-    out = []
-    for event, cfg in (on or {}).items():
-        cfg = cfg or {}
-        if not isinstance(cfg, dict):
-            cfg = {}
-        out.append(Trigger(
-            event=str(event),
-            branches=as_list(cfg.get("branches")),
-            tags=as_list(cfg.get("tags")),
-            paths=as_list(cfg.get("paths")),
-            paths_ignore=as_list(cfg.get("paths-ignore")),
-        ))
-    return out
-
-
-def push_token(job: dict, step: dict) -> str:
-    """Which secret authenticates a push from this step."""
-    refs: list[str] = []
-    for s in job.get("steps") or []:
-        if str(s.get("uses", "")).startswith("actions/checkout"):
-            refs.append(str((s.get("with") or {}).get("token", "")))
-    for scope in (job.get("env") or {}, step.get("env") or {}):
-        for k, v in scope.items():
-            if k in ("GH_TOKEN", "GITHUB_TOKEN", "GIT_TOKEN"):
-                refs.append(str(v))
-    for r in refs:
-        m = re.search(r"secrets\.([A-Za-z0-9_]+)", r)
-        if m and m.group(1) != "GITHUB_TOKEN":
-            return m.group(1)
-    return "GITHUB_TOKEN"
-
-
-def pushed_files(commands: list[str]) -> list[str]:
-    files: list[str] = []
-    for c in commands:
-        if re.match(r"^git\s+add\b", c):
-            args = [a for a in c.split()[2:] if not a.startswith("-")]
-            if re.search(r"\s(-A|--all)\b", c) or "." in args or not args:
-                return ["**"]
-            files.extend(args)
-        elif re.match(r"^git\s+commit\b", c) and re.search(r"\s-a[m]?\b|\s--all\b", c):
-            return ["**"]
-    return files or ["**"]
-
-
-def load_workflow(path: Path, packages: dict[str, str]) -> Workflow:
-    data = yaml.safe_load(path.read_text(encoding="utf-8"))
-    if not isinstance(data, dict):
-        raise ValueError(f"{path.name}: not a mapping")
-    on = data.get("on", data.get(True))  # PyYAML reads a bare `on:` as True
-    triggers = parse_triggers(on)
-    jobs: dict[str, Job] = {}
-    for jname, jdef in (data.get("jobs") or {}).items():
-        jdef = jdef or {}
-        gates: list[Gate] = []
-        pushes: list[Push] = []
-        pr_range = False
-        steps = jdef.get("steps") or []
-        for i, step in enumerate(steps):
-            step = step or {}
-            name = str(step.get("name") or step.get("uses") or step.get("id") or f"step {i}")
-            blob = " ".join(str(v) for v in (step.get("env") or {}).values())
-            if "github.event.pull_request" in blob + str(step.get("run") or ""):
-                pr_range = True
-            run = step.get("run")
-            if run is None:
-                uses = str(step.get("uses") or "")
-                if PUSH_ACTIONS.search(uses):
-                    pushes.append(Push(path.name, jname, i, name, ["**", HISTORY],
-                                       push_token(jdef, step), uses))
-                continue
-            body = str(run)
-            cmds = shell_commands(body)
-            for c in cmds:
-                if re.match(r"^git\s+push\b", c):
-                    # A push lands its files and one new commit on the branch.
-                    pushes.append(Push(path.name, jname, i, name,
-                                       pushed_files(cmds) + [HISTORY],
-                                       push_token(jdef, step), c))
-            gcmds = gate_commands(body)
-            inputs = inputs_for(gcmds, packages)
-            if any(re.match(r"^git\s+log\b", c) for c in cmds):
-                # Reads commit metadata: author, trailers, message.
-                inputs = (inputs or []) + [HISTORY]
-            if not gcmds and EXPLICIT_FAIL.search(body):
-                gcmds = ["(inline verdict: `exit 1` in the step body)"]
-            if gcmds:
-                gates.append(Gate(path.name, jname, i, name, gcmds, inputs, step.get("if")))
-        job = Job(path.name, jname, jdef.get("if"), as_list(jdef.get("needs")) or [],
-                  gates, pushes, pr_range)
-        job.runs_on = {e: runs_on(job.condition, e) for e in ALL_EVENTS}
-        jobs[jname] = job
-    return Workflow(path.name, triggers, jobs)
-
-
-# ── Reachability ─────────────────────────────────────────────────────────────
-
-
-def job_reach(wf: Workflow, job: Job) -> list[str]:
-    """Main-changing events (or the tag push) that actually run this job."""
-    if wf.tag_only:
-        return ["push(tags)"] if job.runs_on["push"] is not False else []
-    return [e for e in wf.main_events() if job.runs_on[e] is not False]
-
-
-def is_manual_only(wf: Workflow, job: Job) -> bool:
-    if wf.trigger("workflow_dispatch") is None or job.runs_on["workflow_dispatch"] is False:
-        return False
-    return all(job.runs_on[e] is False for e in ALL_EVENTS if e != "workflow_dispatch")
-
-
-def is_pr_scoped(wf: Workflow, job: Job) -> bool:
-    if wf.trigger("pull_request") is None or job.runs_on["pull_request"] is False:
-        return False
-    return all(job.runs_on[e] is False for e in MAIN_EVENTS)
-
-
-def gate_reach(wf: Workflow, job: Job, gate: Gate) -> list[str]:
-    return [e for e in job_reach(wf, job) if runs_on(gate.condition, e.split("(")[0]) is not False]
-
-
-def on_path_gates(wf: Workflow, job: Job, push: Push) -> list[Gate]:
-    """Gates that run before `push`: earlier in its job, or in jobs it needs."""
-    out = [g for g in job.gates if g.index < push.index]
-    seen: set[str] = set()
-    todo = list(job.needs)
-    while todo:
-        n = todo.pop()
-        if n in seen or n not in wf.jobs:
-            continue
-        seen.add(n)
-        out.extend(wf.jobs[n].gates)
-        todo.extend(wf.jobs[n].needs)
-    return out
-
-
-def check(wfs: list[Workflow], strict: bool) -> tuple[list[str], list[str]]:
-    found: list[tuple[str, str]] = []  # (key, message)
-    notes: list[str] = []
-
-    def finding(key: str, message: str) -> None:
-        found.append((key, message))
-    all_gates = [g for wf in wfs for j in wf.jobs.values() for g in j.gates]
-
-    # R1 — every gate-bearing job runs on push-to-main or schedule.
-    for wf in wfs:
-        gate_jobs = [j for j in wf.jobs.values() if j.gates]
-        if not gate_jobs:
-            continue
-        reachable = {j.name for j in gate_jobs if job_reach(wf, j)}
-        events = ", ".join(t.event for t in wf.triggers)
-        if not reachable:
-            what = "; ".join(f"{j.name} ({len(j.gates)} gate step(s))" for j in gate_jobs)
-            hint = ""
-            if any(j.pr_range for j in gate_jobs):
-                hint = (" A schedule would not help either: the job reads a "
-                        "github.event.pull_request base..head range, so on any other "
-                        "event it has no range to check — a push trigger needs "
-                        "github.event.before..github.event.after.")
-            finding(
-                f"unreachable:{wf.file}",
-                f"UNREACHABLE  {wf.file}: triggers on [{events}] only; none of its "
-                f"gate-bearing jobs runs on a push to {DEFAULT_BRANCH}, so a commit "
-                f"landing there by merge, direct push, or another workflow's push is "
-                f"never checked. Jobs: {what}.{hint}"
-            )
-            continue
-        for j in gate_jobs:
-            if job_reach(wf, j):
-                continue
-            key = f"{wf.file}:{j.name}"
-            if is_manual_only(wf, j):
-                notes.append(f"MANUAL       {key}: runs on workflow_dispatch only "
-                             f"(if: {j.condition!r}); not a gate on any event")
-            elif is_pr_scoped(wf, j):
-                notes.append(f"PR-SCOPED    {key}: if: {j.condition!r}; {DEFAULT_BRANCH} "
-                             f"is covered by {', '.join(sorted(reachable))}")
-            else:
-                finding(
-                    f"unreachable:{key}",
-                    f"UNREACHABLE  {key}: its `if: {j.condition}` excludes every event "
-                    f"that fires when {DEFAULT_BRANCH} changes ([{events}])."
-                )
-
-    # R2 — a paths-filtered push trigger must cover each reachable gate's inputs.
-    for wf in wfs:
-        push = wf.trigger("push")
-        if push is None or wf.tag_only or (push.paths is None and push.paths_ignore is None):
-            continue
-        for j in wf.jobs.values():
-            for g in j.gates:
-                reach = gate_reach(wf, j, g)
-                if "push" not in reach or "schedule" in reach or not g.inputs:
-                    continue
-                # Commit history is not a path a `paths:` filter can exclude:
-                # a gate over the commit a job itself creates (benchmarks.yml
-                # grading its own push) is R3's business, and a gate over
-                # every push to main (dco.yml) has no filter to gap.
-                gaps = [i for i in g.inputs if i != HISTORY and not covered_by_filter(i, push)]
-                if not gaps:
-                    continue
-                same = [o for o in all_gates if o is not g
-                        and {normalise(c) for c in o.commands} & {normalise(c) for c in g.commands}]
-                elsewhere = []
-                for o in same:
-                    owf = next(w for w in wfs if w.file == o.workflow)
-                    oreach = gate_reach(owf, owf.jobs[o.job], o)
-                    opush = owf.trigger("push")
-                    if "schedule" in oreach or ("push" in oreach and opush is not None
-                                                and all(covered_by_filter(i, opush) for i in gaps)):
-                        elsewhere.append(o.where)
-                if elsewhere:
-                    notes.append(f"FILTER-GAP   {g.where}: reads {', '.join(gaps)} outside "
-                                 f"{wf.file}'s push paths; covered by {'; '.join(elsewhere)}")
-                else:
-                    finding(
-                        f"filter-gap:{wf.file}:{j.name}:step[{g.index}]",
-                        f"FILTER-GAP   {g.where}: reads {', '.join(gaps)}, but {wf.file}'s push "
-                        f"trigger fires only for paths [{', '.join(push.paths or [])}]"
-                        + (f" minus [{', '.join(push.paths_ignore)}]" if push.paths_ignore else "")
-                        + f". A push to {DEFAULT_BRANCH} changing only those files never runs "
-                        f"it, and no other {DEFAULT_BRANCH}-reachable job runs the same command."
-                    )
-
-    # R3 — a GITHUB_TOKEN push creates no events; the gates must already have run.
-    # Readers are the gates a push to main would have triggered: tag-only
-    # workflows see the tree only when a tag is cut, so they are not among them.
-    tag_only = {wf.file for wf in wfs if wf.tag_only}
-    branch_gates = [g for g in all_gates if g.workflow not in tag_only]
-    for wf in wfs:
-        for j in wf.jobs.values():
-            for p in j.pushes:
-                if p.creates_events:
-                    notes.append(f"BOT-PUSH     {p.where}: pushes with secrets.{p.token}, which "
-                                 f"does create workflow runs; covered by the push rules above")
-                    continue
-                before = on_path_gates(wf, j, p)
-                for f in p.files:
-                    on_path = [g for g in before if g.reads(f)]
-                    readers = [g for g in branch_gates if g.reads(f)]
-                    bypassed = by_job([g for g in readers if g not in on_path])
-                    what = "the commit it adds" if f == HISTORY else f
-                    if not readers:
-                        notes.append(f"BOT-PUSH     {p.where}: pushes {what} with GITHUB_TOKEN "
-                                     f"(no workflow run follows); no known gate reads it")
-                        continue
-                    if not on_path:
-                        finding(
-                            f"bypassed:{wf.file}:{j.name}:step[{p.index}]:{f}",
-                            f"BYPASSED     {p.where}: pushes {what} to {DEFAULT_BRANCH} with "
-                            f"GITHUB_TOKEN, which creates no workflow run, and no gate that reads "
-                            f"{what} runs earlier in that job. Gates it escapes: {bypassed}"
-                        )
-                        continue
-                    notes.append(f"BOT-PUSH     {p.where}: pushes {what}; checked first by "
-                                 + "; ".join(g.where for g in on_path))
-                    if bypassed:
-                        line = (f"{what} is also read by gates that do not run before "
-                                f"{p.where} pushes it: {bypassed}")
-                        if strict:
-                            finding(f"bypassed:{wf.file}:{j.name}:step[{p.index}]:{f}:strict",
-                                    f"BYPASSED     {line}")
-                        else:
-                            notes.append(f"NOT-ON-PATH  {line}")
-
-    findings: list[str] = []
-    for key, message in found:
-        if key in ALLOWED:
-            notes.append(f"ALLOWED      {key}: {ALLOWED[key]}")
-        else:
-            findings.append(f"{message}\n               key: {key}")
-    return findings, notes
-
-
-def by_job(gates: list[Gate]) -> str:
-    """`ci.yml:fmt (2 steps); ci.yml:test (14 steps)` — readable at 40 gates."""
-    groups: dict[str, list[Gate]] = {}
-    for g in gates:
-        groups.setdefault(f"{g.workflow}:{g.job}", []).append(g)
-    parts = []
-    for key, gs in groups.items():
-        parts.append(f"{key} {gs[0].name!r}" if len(gs) == 1 else f"{key} ({len(gs)} steps)")
-    return "; ".join(parts)
-
-
-# ── Explain ──────────────────────────────────────────────────────────────────
-
-
-def explain(wfs: list[Workflow]) -> None:
-    for wf in wfs:
-        print(f"\n{wf.file}")
-        for t in wf.triggers:
-            print(f"  on: {t.describe()}")
-        me = wf.main_events()
-        print(f"  fires when {DEFAULT_BRANCH} changes via: "
-              + (", ".join(me) if me else ("push(tags) only" if wf.tag_only else "nothing")))
-        for j in wf.jobs.values():
-            if not j.gates and not j.pushes:
-                continue
-            reach = job_reach(wf, j)
-            cond = f" if: {j.condition!r}" if j.condition is not None else ""
-            unknown = [e for e, v in j.runs_on.items() if v is None]
-            tag = ", ".join(reach) if reach else (
-                "manual only" if is_manual_only(wf, j) else
-                "PR-scoped" if is_pr_scoped(wf, j) else "UNREACHABLE")
-            print(f"  job {j.name}{cond}"
-                  + (f" (needs {', '.join(j.needs)})" if j.needs else "")
-                  + f" -> reachable on {DEFAULT_BRANCH} via: {tag}"
-                  + (f" [if: not understood for {', '.join(unknown)}; assumed to run]" if unknown else "")
-                  + (" [reads a pull_request base..head range]" if j.pr_range else ""))
-            for g in j.gates:
-                ins = "inputs unknown" if g.inputs is None else (
-                    "no tracked inputs" if not g.inputs else "reads " + ", ".join(g.inputs))
-                print(f"    gate step[{g.index}] {g.name!r}: {ins}")
-                for c in g.commands:
-                    print(f"        $ {c[:110]}")
-            for p in j.pushes:
-                print(f"    PUSH step[{p.index}] {p.name!r}: `{p.line}` with "
-                      f"{'secrets.' + p.token if p.creates_events else 'GITHUB_TOKEN (creates no workflow run)'}"
-                      f"; changes {', '.join(p.files)}")
-                before = on_path_gates(wf, j, p)
-                for f in p.files:
-                    on = [g for g in before if g.reads(f)]
-                    print(f"        {f}: checked before the push by "
-                          + ("; ".join(g.where for g in on) if on else "NOTHING"))
-
-
-# ── Main ─────────────────────────────────────────────────────────────────────
+# The package sits beside this script in scripts/lib/, which is not on
+# sys.path. Resolving the script's own location means the same invocation
+# works from the repository root and from any other cwd.
+sys.path.insert(0, str(Path(__file__).resolve().parent / "lib"))
+
+from gate_reachability.inputs import package_dirs  # noqa: E402
+from gate_reachability.model import WORKFLOWS, Workflow  # noqa: E402
+from gate_reachability.report import explain, report  # noqa: E402
+from gate_reachability.rules import check  # noqa: E402
+from gate_reachability.workflow import LOAD_ERRORS, load_workflow  # noqa: E402
 
 
 def main() -> int:
@@ -957,7 +98,7 @@ def main() -> int:
     for f in files:
         try:
             wfs.append(load_workflow(f, packages))
-        except (yaml.YAMLError, ValueError, OSError) as e:
+        except LOAD_ERRORS as e:
             print(f"check_gate_reachability: cannot read {f.name}: {e}", file=sys.stderr)
             return 2
 
@@ -966,20 +107,7 @@ def main() -> int:
         print()
 
     findings, notes = check(wfs, args.strict)
-    n_jobs = sum(1 for wf in wfs for j in wf.jobs.values() if j.gates)
-    n_gates = sum(len(j.gates) for wf in wfs for j in wf.jobs.values())
-    n_unknown = sum(1 for wf in wfs for j in wf.jobs.values() for g in j.gates if g.inputs is None)
-    n_push = sum(len(j.pushes) for wf in wfs for j in wf.jobs.values())
-
-    for line in notes:
-        print(f"  {line}")
-    if findings:
-        print("check_gate_reachability: gates that do not run on an event that can break them")
-        for line in findings:
-            print(f"  {line}")
-    print(f"check_gate_reachability: {len(wfs)} workflows, {n_jobs} gate-bearing jobs, "
-          f"{n_gates} gate steps ({n_unknown} with unknown inputs), {n_push} in-workflow "
-          f"push(es), {len(findings)} finding(s)")
+    report(wfs, findings, notes)
     return 1 if findings else 0
